+++ skillfarm/__init__.py
@@ -3,7 +3,7 @@
 # AA Skillfarm
 from skillfarm.app_settings import SKILLFARM_APP_NAME
 
-__version__ = "0.5.4a10"
+__version__ = "0.5.4"
 __title__ = "Skillfarm"
 
 USER_AGENT_TEXT = f"{SKILLFARM_APP_NAME} v{__version__}"